+++ tests/serialize/test_PlySerialize.py
@@ -1,6 +1,4 @@
 import pytest
-import tempfile
-import os
 import numpy as np
 from tests.serialize import TestSerializeBase as base_class
 
@@ -12,59 +10,6 @@ class TestPlyDict:
     def geom_cls(self):
         from yggdrasil.serialize.PlySerialize import PlyDict
         return PlyDict
-
-    def test_mesh(self, geom_cls, geom_dict):
-        r"""Test construction from a mesh."""
-        x = geom_cls(geom_dict)
-        mesh = x.mesh
-        y = geom_cls.from_mesh(mesh)
-        assert x.mesh == y.mesh
-        assert x.nvert != y.nvert
-        assert x.nface == y.nface
-        z = geom_cls.from_mesh(mesh, prune_duplicates=True)
-        assert x.mesh == z.mesh
-        assert x.nvert == z.nvert
-        assert x.nface == z.nface
-
-    def test_mesh_structured(self, geom_cls, geom_dict):
-        r"""Test construction from a numpy structured array."""
-        from numpy.lib.recfunctions import unstructured_to_structured
-        x = geom_cls(geom_dict)
-        field_names = ['x1', 'y1', 'z1',
-                       'x2', 'y2', 'z2',
-                       'x3', 'y3', 'z3']
-        formats = ['f8' for _ in field_names]
-        dtype = np.dtype(dict(names=field_names, formats=formats))
-        mesh = unstructured_to_structured(np.array(x.mesh), dtype=dtype)
-        y = geom_cls.from_mesh(mesh)
-        assert x.mesh == y.mesh
-        assert x.nvert != y.nvert
-        assert x.nface == y.nface
-        z = geom_cls.from_mesh(mesh, prune_duplicates=True)
-        assert x.mesh == z.mesh
-        assert x.nvert == z.nvert
-        assert x.nface == z.nface
-
-    def test_mesh_file(self, geom_cls, geom_dict):
-        r"""Test construciton from a mesh file."""
-        x = geom_cls(geom_dict)
-        mesh = np.asarray(x.mesh)
-        ftemp = tempfile.NamedTemporaryFile(delete=False)
-        np.savetxt(ftemp, mesh)
-        fname = ftemp.name
-        ftemp.close()
-        try:
-            y = geom_cls.from_mesh(fname)
-            assert x.mesh == y.mesh
-            assert x.nvert != y.nvert
-            assert x.nface == y.nface
-            z = geom_cls.from_mesh(fname, prune_duplicates=True)
-            assert x.mesh == z.mesh
-            assert x.nvert == z.nvert
-            assert x.nface == z.nface
-        finally:
-            if os.path.isfile(fname):
-                os.remove(fname)
 
 
 class TestPlySerialize(base_class):
@@ -90,19 +35,7 @@ class TestPlySerialize(base_class):
         for x in testing_options['objects']:
             scalar_arr = 10 * np.arange(x.count_elements('faces')).astype(
                 'float')
-            with pytest.raises(NotImplementedError):
-                x.apply_scalar_map(scalar_arr, scale_by_area=True)
-            new_faces = []
-            if 'Obj' in class_name:
-                for f in x['faces']:
-                    if len(f) == 3:
-                        new_faces.append(f)
-            else:
-                for f in x['faces']:
-                    if len(f['vertex_index']) == 3:
-                        new_faces.append(f)
             odict = x.as_dict()
-            odict['face'] = new_faces
             for scale in ['linear', 'log']:
                 ox = type(x)(odict)
                 o2 = type(x)(odict)