+++ yggdrasil/languages/C/datatypes/serialization.h
@@ -11,6 +11,7 @@
 #include "utils.h"
 
 #define RAPIDJSON_YGGDRASIL
+#define RAPIDJSON_HAS_STDSTRING 1
 #include "rapidjson/document.h"
 #include "rapidjson/writer.h"
 #include "rapidjson/prettywriter.h"